requirements.txt
fastapi>=0.92.0
loguru>=0.6.0
openai>=0.27.0
Pillow>=9.4.0
pydantic>=1.10.5
pytest>=7.2.2
python-dotenv>=1.0.0
requests>=2.28.2
uvicorn>=0.20.0
httpx